requirements.txt
torch
pydub
openai-whisper
numpy
json
spleeter
ffmpeg